src/main/java/com/equipe1/service/UserService.java
package com.equipe1.service;

import com.equipe1.model.User;
import com.equipe1.model.Rappel;
import com.equipe1.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.stereotype.Service;
import org.springframework.web.client.HttpServerErrorException;
import org.springframework.web.server.ResponseStatusException;

import java.util.List;
import java.util.Optional;

@Service
public class UserService {

    @Autowired
    private UserRepository userRepository;

    public List<User> getAllUsers() {
        return userRepository.findAll();
    }

    public User getUser(String email, String password) {
        Optional<User> optionalUser = userRepository.findByEmail(email);
        if (optionalUser.isEmpty())
            return null;
        User user = optionalUser.get();
        return user.getPassword().equals(password) ? user: null;
    }

    public User getUserByEmail(String email) {
        Optional<User> optionalUser = userRepository.findByEmail(email);
        if (optionalUser.isEmpty())
            return null;
        User user = optionalUser.get();
        return user.getEmail().equals(email) ? user: null;
    }

    public User getById(long id) throws ResponseStatusException {
        return userRepository.findById(id).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "no user with id " + id));
    }
}